ai_service/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
pydantic==2.9.2
PyJWT==2.9.0
